// A1/frontend/src/models/User.ts
export interface Users{
    user_id?: number;
    username: string;
    bio: string;
    email: string;
    birthday: string;
    country: string;
    gender: string;
    nr_of_dogs?: number;
    nr_of_owners?: number;
    nr_of_toys?: number;
    nr_of_dogowners?: number;
    page_size?: number;
}